virtlogd: use %llu to print 64bit types
Otherwise we fail on 32bit with:

CC     logging/virtlogd-log_daemon_dispatch.o
logging/log_daemon_dispatch.c: In function 'virLogManagerProtocolDispatchDomainReadLogFile':
logging/log_daemon_dispatch.c:120:9: error: format '%zu' expects argument of type 'size_t', but argument 7 has type 'uint64_t' [-Werror=format]

+++ src/logging/log_daemon_dispatch.c
@@ -118,8 +118,9 @@ virLogManagerProtocolDispatchDomainReadLogFile(virNetServerPtr server ATTRIBUTE_
 
     if (args->maxlen > VIR_LOG_MANAGER_PROTOCOL_STRING_MAX) {
         virReportError(VIR_ERR_INTERNAL_ERROR,
-                       _("Requested data len %zu is larger than maximum %d"),
-                       args->maxlen, VIR_LOG_MANAGER_PROTOCOL_STRING_MAX);
+                       _("Requested data len %llu is larger than maximum %d"),
+                       (unsigned long long)args->maxlen,
+                       VIR_LOG_MANAGER_PROTOCOL_STRING_MAX);
         goto cleanup;
     }
